requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.9.1
torchmetrics==0.11.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
sh            # for running bash commands in some tests (linux/macos only)
scipy
matplotlib      # to visualize images in noteoobks

# h5py==3.7.0
# setuptools==59.5.0
# pycocotools==2.0.6